ai/fonio-forecast/requirements.txt
grpcio>=1.14.2
grpcio-tools>=1.14.1
cntk-gpu==2.6
pandas==0.23.4